edit pass on existing samples
