+++ docs/csharp/language-reference/keywords/snippets/DefineTypes.cs
@@ -1,0 +1,61 @@
+﻿using System;
+
+namespace keywords;
+//<snippet15>
+interface IPoint
+{
+    // Property signatures:
+    int X
+    {
+        get;
+        set;
+    }
+
+    int Y
+    {
+        get;
+        set;
+    }
+
+    double Distance
+    {
+        get;
+    }
+}
+
+class Point : IPoint
+{
+    // Constructor:
+    public Point(int x, int y)
+    {
+        X = x;
+        Y = y;
+    }
+
+    // Property implementation:
+    public int X { get; set; }
+
+    public int Y { get; set; }
+
+    // Property implementation
+    public double Distance =>
+       Math.Sqrt(X * X + Y * Y);
+
+}
+
+class MainClass
+{
+    static void PrintPoint(IPoint p)
+    {
+        Console.WriteLine("x={0}, y={1}", p.X, p.Y);
+    }
+
+    static void Main()
+    {
+        IPoint p = new Point(2, 3);
+        Console.Write("My Point: ");
+        PrintPoint(p);
+    }
+}
+// Output: My Point: x=2, y=3
+//</snippet15>
